When trying to deploy the app to my own heroku instance, I see the following error in the logs:
ActionView::Template::Error (The asset "application.css" is not present in the asset pipeline.

Upon seeing this error, I googled the message and came across this link
- https://www.reddit.com/r/rails/comments/71by4m/help_the_asset_applicationcss_is_not_present_in/

Therefore, in an attempt to fix this issue, I'm precompiling our
application.css asset with this change. If it does not work, I will
revert this change.