Sets yarn to specific version - 1.22.17 for now.
When opening PRs or deploying to Heroku, I noticed a problem when the
build tried to install yarn. Here is an example:

https://github.com/rubyforgood/shelter-assist/runs/6885776407?check_suite_focus=true

There seems to be an issue with a yarn version mismatch. I *believe*
yarn-actions is just getting the latest version of yarn that it can
find, which may be different from the version of yarn specified in our
package.json.

Running the `yarn set version` command sets yarn to a particular version
and adds the files in this commit. I think as we upgrade yarn in the
future, we can re-run `yarn set version` to update the .yarnrc files.